+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All the changes, additions, bugfixes, etc. made to this project.
   - Deploy section
 - Add `box-sizing: border-box;` so it doesn't break whenever a padding is added, and the full width gets calculated WITH the border and not FROM.
 - Add placeholder color in Task form
+- Implement `import` reducer to change the state with a given raw JSON string
 
 ### Modified
 - Redesign a little bit the task form in `<AddTask>`